requirements.txt
load-dotenv==0.1.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3

